Refactored docs to fetch from a ref
