--- docs/structure.json
@@ -1,23 +1,0 @@
-{
-  "Getting Started": [
-    ["installation.mdx", "Installation"],
-    ["typescript.mdx", "TypeScript"],
-    ["thinking-in-effection.mdx", "Thinking in Effection"],
-    ["async-rosetta-stone.mdx", "Async Rosetta Stone"],
-    ["tutorial.mdx", "Tutorial"]
-  ],
-  "Learn Effection": [
-    ["operations.mdx", "Operations"],
-    ["actions.mdx", "Actions and Suspensions"],
-    ["resources.mdx", "Resources"],
-    ["spawn.mdx", "Spawn"],
-    ["collections.mdx", "Streams and Subscriptions"],
-    ["events.mdx", "Events"],
-    ["errors.mdx", "Error Handling"],
-    ["context.mdx", "Context"]
-  ],
-  "Advanced": [
-    ["scope.mdx", "Scope"],
-    ["processes.mdx", "Processes"]
-  ]
-}
